working bounds propagator + moved the propagator to int_all_different.rs to comply with the latest changes in huub
